refactor: centralize knex client
