+++ db/client.js
@@ -1,0 +1,4 @@
+import knex from 'knex';
+import config from './knexfile.js';
+const db = knex(config);
+export default db;
